Split the existing file into two
